feat: add backup treigger on player quit
When the interval task is enabled and `skip_if_no_players` is set, a
backup will now be triggered when the last player quits. This handles
the case where no backup was performed after the player left, and the
latest backup remained outdated.

Additionally, a new option `backup.on_player_quit` is added to trigger a
backup whenever a player quits the server.